requirements.txt
Mastodon.py
html2text
requests